🪒 Razor: Unify duplicated CircuitBreaker implementations
Unified the nearly identical `CircuitBreaker` implementations found in `src/storage/sharding/network.rs` and `src/index/vector/distributed.rs` into a single, cohesive module at `src/core/circuit_breaker.rs`. Re-exported the relevant symbols back into the original modules to maintain API compatibility while significantly reducing boilerplate and duplicate logic. Evaluated by code review as correct and robust.
