+++ src/storage/sharding/network.rs
@@ -7,13 +7,14 @@
 #![allow(clippy::collapsible_if)]
 
 use super::types::{ShardId, ShardState};
+pub use crate::core::circuit_breaker::{CircuitBreaker, CircuitBreakerConfig, CircuitState};
 use crate::core::hlc::HybridTimestamp;
 use crate::core::id::{EdgeId, NodeId, TxId};
 use std::collections::HashMap;
 use std::fmt;
 use std::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
 use std::sync::{Arc, RwLock};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 /// Error types for network operations.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -248,223 +249,6 @@ pub trait ShardClient: Send + Sync + fmt::Debug {
 
     /// Perform a health check.
     fn health_check(&self) -> NetworkResult<()>;
-}
-
-/// Circuit breaker state.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CircuitState {
-    /// Circuit is closed (normal operation).
-    Closed,
-    /// Circuit is open (rejecting requests).
-    Open,
-    /// Circuit is half-open (testing if service recovered).
-    HalfOpen,
-}
-
-/// Configuration for circuit breaker.
-#[derive(Debug, Clone)]
-pub struct CircuitBreakerConfig {
-    /// Number of failures before opening circuit.
-    pub failure_threshold: usize,
-    /// Duration to keep circuit open.
-    pub open_duration: Duration,
-    /// Number of successes in half-open to close circuit.
-    pub success_threshold: usize,
-    /// Window for counting failures.
-    pub failure_window: Duration,
-}
-
-impl Default for CircuitBreakerConfig {
-    fn default() -> Self {
-        Self {
-            failure_threshold: 5,
-            open_duration: Duration::from_secs(30),
-            success_threshold: 3,
-            failure_window: Duration::from_secs(60),
-        }
-    }
-}
-
-/// Circuit breaker for a single shard connection.
-#[derive(Debug)]
-pub struct CircuitBreaker {
-    config: CircuitBreakerConfig,
-    state: RwLock<CircuitState>,
-    failure_count: AtomicUsize,
-    success_count: AtomicUsize,
-    last_failure: RwLock<Option<Instant>>,
-    opened_at: RwLock<Option<Instant>>,
-}
-
-impl CircuitBreaker {
-    /// Create a new circuit breaker.
-    pub fn new(config: CircuitBreakerConfig) -> Self {
-        Self {
-            config,
-            state: RwLock::new(CircuitState::Closed),
-            failure_count: AtomicUsize::new(0),
-            success_count: AtomicUsize::new(0),
-            last_failure: RwLock::new(None),
-            opened_at: RwLock::new(None),
-        }
-    }
-
-    /// Get the current state.
-    ///
-    /// If the lock is poisoned, returns Closed (fail-open) to prevent
-    /// cascading failures.
-    pub fn state(&self) -> CircuitState {
-        self.maybe_transition();
-        self.state
-            .read()
-            .map(|s| *s)
-            .unwrap_or(CircuitState::Closed)
-    }
-
-    /// Check if requests should be allowed.
-    ///
-    /// If the lock is poisoned, returns true (fail-open) to prevent
-    /// cascading failures.
-    pub fn should_allow(&self) -> bool {
-        self.maybe_transition();
-        let state = self
-            .state
-            .read()
-            .map(|s| *s)
-            .unwrap_or(CircuitState::Closed);
-        matches!(state, CircuitState::Closed | CircuitState::HalfOpen)
-    }
-
-    /// Record a successful request.
-    pub fn record_success(&self) {
-        let state = match self.state.read() {
-            Ok(s) => *s,
-            Err(_) => return, // Lock poisoned, silently skip
-        };
-
-        match state {
-            CircuitState::Closed => {
-                // Reset failure count on success
-                self.failure_count.store(0, Ordering::SeqCst);
-            }
-            CircuitState::HalfOpen => {
-                let successes = self.success_count.fetch_add(1, Ordering::SeqCst) + 1;
-                if successes >= self.config.success_threshold {
-                    // Transition to closed
-                    if let Ok(mut s) = self.state.write() {
-                        *s = CircuitState::Closed;
-                    }
-                    self.failure_count.store(0, Ordering::SeqCst);
-                    self.success_count.store(0, Ordering::SeqCst);
-                }
-            }
-            CircuitState::Open => {
-                // Shouldn't happen, but ignore
-            }
-        }
-    }
-
-    /// Record a failed request.
-    pub fn record_failure(&self) {
-        let state = match self.state.read() {
-            Ok(s) => *s,
-            Err(_) => return, // Lock poisoned, silently skip
-        };
-
-        match state {
-            CircuitState::Closed => {
-                // Check if we should reset due to window expiry
-                if let Ok(last) = self.last_failure.read() {
-                    if let Some(last_time) = *last {
-                        if last_time.elapsed() > self.config.failure_window {
-                            self.failure_count.store(0, Ordering::SeqCst);
-                        }
-                    }
-                }
-
-                let failures = self.failure_count.fetch_add(1, Ordering::SeqCst) + 1;
-
-                if let Ok(mut last) = self.last_failure.write() {
-                    *last = Some(Instant::now());
-                }
-
-                if failures >= self.config.failure_threshold {
-                    // Transition to open
-                    if let Ok(mut s) = self.state.write() {
-                        *s = CircuitState::Open;
-                    }
-                    if let Ok(mut opened) = self.opened_at.write() {
-                        *opened = Some(Instant::now());
-                    }
-                }
-            }
-            CircuitState::HalfOpen => {
-                // Any failure in half-open goes back to open
-                if let Ok(mut s) = self.state.write() {
-                    *s = CircuitState::Open;
-                }
-                if let Ok(mut opened) = self.opened_at.write() {
-                    *opened = Some(Instant::now());
-                }
-                self.success_count.store(0, Ordering::SeqCst);
-            }
-            CircuitState::Open => {
-                // Already open, update opened_at
-                if let Ok(mut opened) = self.opened_at.write() {
-                    *opened = Some(Instant::now());
-                }
-            }
-        }
-    }
-
-    /// Check and perform state transitions based on time.
-    fn maybe_transition(&self) {
-        let state = match self.state.read() {
-            Ok(s) => *s,
-            Err(_) => return, // Lock poisoned, skip transition
-        };
-
-        if state == CircuitState::Open {
-            if let Ok(opened) = self.opened_at.read() {
-                if let Some(opened_time) = *opened {
-                    if opened_time.elapsed() >= self.config.open_duration {
-                        // Transition to half-open
-                        if let Ok(mut s) = self.state.write() {
-                            *s = CircuitState::HalfOpen;
-                        }
-                        self.success_count.store(0, Ordering::SeqCst);
-                    }
-                }
-            }
-        }
-    }
-
-    /// Get remaining time before circuit can close.
-    pub fn remaining_open_time(&self) -> Option<Duration> {
-        let state = self.state.read().ok()?;
-        if *state != CircuitState::Open {
-            return None;
-        }
-
-        if let Ok(opened) = self.opened_at.read() {
-            if let Some(opened_time) = *opened {
-                let elapsed = opened_time.elapsed();
-                if elapsed < self.config.open_duration {
-                    return Some(self.config.open_duration - elapsed);
-                }
-            }
-        }
-        None
-    }
-
-    /// Reset the circuit breaker to closed state.
-    pub fn reset(&self) {
-        if let Ok(mut s) = self.state.write() {
-            *s = CircuitState::Closed;
-        }
-        self.failure_count.store(0, Ordering::SeqCst);
-        self.success_count.store(0, Ordering::SeqCst);
-    }
 }
 
 /// Configuration for connection pool.
